Create 'bytesToDate()' as a module import.
